+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Changes
+
+- Add `namespace_for_prefix` function to look up namespace in the context of a
+  node.
+
 ## [0.14.0] - 2023-08-24
 
 ### Changes
@@ -390,8 +395,8 @@
 Initial public release.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/faassen/xot/compare/v0.14.0...HEAD
 
+[Unreleased]: https://github.com/faassen/xot/compare/v0.14.0...HEAD
 [0.14.0]: https://github.com/faassen/xot/compare/v0.13.5...v0.14.0
 [0.13.5]: https://github.com/faassen/xot/compare/v0.13.4...v0.13.5
 [0.13.4]: https://github.com/faassen/xot/compare/v0.13.3...v0.13.4